move room view to views package
